requirements.txt
streamlit
pandas
matplotlib
python-docx
docx2pdf
